Disbursement flow with grantee
Former-commit-id: ee1509cf5c40a7ad4de58c06c47d2e51a512e896

+++ src/main/java/org/codealpha/gmsservice/controllers/UserController.java
@@ -339,14 +339,14 @@ public class UserController {
         if(activeGrantSummaryCommitted==null){
             return null;
         }
-        Long disbursedAmount = dashboardService.getActiveGrantDisbursedAmountForGranter(tenantOrg.getId(),status);
+        Double disbursedAmount = dashboardService.getActiveGrantDisbursedAmountForGranter(tenantOrg.getId(),status);
         Filter categoryFilter = new Filter();
         categoryFilter.setName(WordUtils.capitalizeFully(status+" Grants"));
         categoryFilter.setTotalGrants(Long.valueOf(activeGrantSummaryCommitted.getGrantCount()));
         SimpleDateFormat sd = new SimpleDateFormat("yyyy");
         categoryFilter.setPeriod(sd.format(activeGrantSummaryCommitted.getPeriodStart()) + "-" + sd.format(activeGrantSummaryCommitted.getPeriodEnd()));
         categoryFilter.setCommittedAmount(Long.valueOf(activeGrantSummaryCommitted.getCommittedAmount()));
-        categoryFilter.setDisbursedAmount(Long.valueOf(disbursedAmount));
+        categoryFilter.setDisbursedAmount(disbursedAmount.longValue());
         List<GranterReportStatus> reportStatuses = null;
         List<DetailedSummary> reportSummaryList = new ArrayList<>();
         if(status.equalsIgnoreCase("ACTIVE")) {
